% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FairnessTest.R
\name{FairnessTest}
\alias{FairnessTest}
\title{Fairness testing of an ML model}
\description{
Fairness testing of an ML model

Fairness testing of an ML model
}
\examples{
if (require("randomForest")) {
  # data pre-processing 
  compas <- fairml::compas
  compas <- compas \%>\% drop_na()
  compas <- compas \%>\% distinct()
  # Train a model
  set.seed(142)
  rf = randomForest(two_year_recid ~ ., data = compas[-17L, ])
  # Create a predictor object
  predictor = iml::Predictor$new(rf, type = "prob")
  # Find differences of the prediction of counterfactuals and original instance 
  fairness_obj = FairnessTest$new(predictor, df = compas, sensitive_attribute = "race", n_generations = 175)
  # Print the results
  difference = fairness_obj$get_prediction_difference(x_interest, desired_level = "Caucasian", desired_prob = c(0.5,1))
  print(difference)
}


}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{predictor}}{(\link[iml]{Predictor})\cr
The object (created with \code{iml::Predictor$new()}) holding the machine learning model and the data.}

\item{\code{df}}{(dataframe)\cr
The dataset or dataframe for the prediction}

\item{\code{sensitive_attribute}}{(\code{string})\cr
The name of sensitive_attribute.}

\item{\code{n_generations}}{(\code{integerish(1)})\cr
The number of generations. Default is \code{175L}.}

\item{\code{cfs}}{(dataframe)\cr
A datframe consisting counterfactuals (plausible)}

\item{\code{org_cfs}}{(dataframe)\cr
A dataframe containing all counterfactuals}

\item{\code{pred_diff}}{(dataframe)\cr
A dataframe consisting the prediction differences}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FairnessTest$new()}}
\item \href{#method-generate_counterfactuals}{\code{FairnessTest$generate_counterfactuals()}}
\item \href{#method-plot_tSNE}{\code{FairnessTest$plot_tSNE()}}
\item \href{#method-print_prediction}{\code{FairnessTest$print_prediction()}}
\item \href{#method-print_counterfactuals}{\code{FairnessTest$print_counterfactuals()}}
\item \href{#method-prediction_percentages}{\code{FairnessTest$prediction_percentages()}}
\item \href{#method-get_prediction_difference_mean}{\code{FairnessTest$get_prediction_difference_mean()}}
\item \href{#method-get_cfactuals_count}{\code{FairnessTest$get_cfactuals_count()}}
\item \href{#method-clone}{\code{FairnessTest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new FairnessTest object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessTest$new(
  predictor = NULL,
  df = NULL,
  sensitive_attribute = NULL,
  n_generations = 175L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictor}}{(\link[iml]{Predictor})\cr
The object (created with \code{iml::Predictor$new()}) holding the machine learning model and the data.}

\item{\code{df}}{(dataframe)\cr
The dataset or dataframe for the prediction}

\item{\code{sensitive_attribute}}{(\code{string})\cr
The name of sensitive_attribute.}

\item{\code{n_generations}}{(\code{integerish(1)})\cr
The number of generations. Default is \code{175L}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_counterfactuals"></a>}}
\if{latex}{\out{\hypertarget{method-generate_counterfactuals}{}}}
\subsection{Method \code{generate_counterfactuals()}}{
Creates a dataframe with the differences of predictions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessTest$generate_counterfactuals(
  x_interest,
  desired_level,
  desired_prob,
  fixed_features = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_interest}}{(\code{data.table(1)} | \code{data.frame(1)}) \cr
A single row with the observation of interest.}

\item{\code{desired_level}}{(\code{character(1)} | \code{NULL}) \cr
The desired class. If \code{NULL} (default) then \code{predictor$class} is taken.}

\item{\code{desired_prob}}{(\code{numeric(1)} | \code{numeric(2)}) \cr
The desired predicted probability of the \code{desired_class}. It can be a numeric scalar or a vector with two
numeric values that specify a probability range.
For hard classification tasks this can be set to \code{0} or \code{1}, respectively.}

\item{\code{fixed_features}}{(\code{character()} | \code{NULL})\cr
Names of features that are not allowed to change. \code{NULL} (default) allows to change all features.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A dataframe with the differences of predictions of the original instance and the counterfactuals
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_tSNE"></a>}}
\if{latex}{\out{\hypertarget{method-plot_tSNE}{}}}
\subsection{Method \code{plot_tSNE()}}{
Creates a dataframe with the differences of predictions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessTest$plot_tSNE(x_interest, factor_variable = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_interest}}{(\code{data.table(1)} | \code{data.frame(1)}) \cr
A single row with the observation of interest.}

\item{\code{factor_variable}}{(\code{string})\cr
Names of features that are not allowed to change}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A dataframe with the differences of predictions of the original instance and the counterfactuals
#' @description
#' Creates a dataframe with the differences of predictions
#' @template x_interest
#' @param desired_level (\code{character(1)} | \code{NULL}) \cr
#'   The desired class. If \code{NULL} (default) then \code{predictor$class} is taken.
get_wapd = function(){
diff = self$pred_diff
n_cf = length(diff)
n_pos = length(diff\link{diff > 0})
n_neg = length(diff\link{diff < 0})
sum_pos = sum(diff\link{diff > 0})
sum_neg = sum(diff\link{diff < 0})

ans = ((n_pos/n_cf)*sum_pos + (n_neg/n_cf)*sum_neg)/n_cf
return(round(ans,2))

},
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_prediction"></a>}}
\if{latex}{\out{\hypertarget{method-print_prediction}{}}}
\subsection{Method \code{print_prediction()}}{
returns a dataframe with the required column for tSNE plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessTest$print_prediction()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A dataframe of counterfactuals with the required column for tSNE plot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_counterfactuals"></a>}}
\if{latex}{\out{\hypertarget{method-print_counterfactuals}{}}}
\subsection{Method \code{print_counterfactuals()}}{
prints the counterfactuals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessTest$print_counterfactuals()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
prints the generated counterfactuals
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prediction_percentages"></a>}}
\if{latex}{\out{\hypertarget{method-prediction_percentages}{}}}
\subsection{Method \code{prediction_percentages()}}{
prints the percentage for the predictions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessTest$prediction_percentages(x_interest)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_interest}}{(\code{data.table(1)} | \code{data.frame(1)}) \cr
A single row with the observation of interest.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the percentage for the predictions of generated counterfactuals
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_prediction_difference_mean"></a>}}
\if{latex}{\out{\hypertarget{method-get_prediction_difference_mean}{}}}
\subsection{Method \code{get_prediction_difference_mean()}}{
returns mean of the prediction of the counterfactuals as the opposite of the x_interest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessTest$get_prediction_difference_mean()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
mean of the prediction of the counterfactuals as the opposite of the x_interest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_cfactuals_count"></a>}}
\if{latex}{\out{\hypertarget{method-get_cfactuals_count}{}}}
\subsection{Method \code{get_cfactuals_count()}}{
Gives the total count of counterfactuals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessTest$get_cfactuals_count()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessTest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
