% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FairnessTest.R
\name{FairnessTest}
\alias{FairnessTest}
\title{Fairness testing of an ML model}
\description{
Fairness testing of an ML model

Fairness testing of an ML model
}
\examples{
if (require("randomForest")) {
  # data pre-processing 
  compas <- fairml::compas
  compas <- compas \%>\% drop_na()
  compas <- compas \%>\% distinct()
  # Train a model
  set.seed(142)
  rf = randomForest(two_year_recid ~ ., data = compas[-17L, ])
  # Create a predictor object
  predictor = iml::Predictor$new(rf, type = "prob")
  # Find differences of the prediction of counterfactuals and original instance 
  fairness_obj = FairnessTest$new(predictor, df = compas, sensitive_attribute = "race", row_num = 17L, n_generations = 175)
  # Print the results
  difference = fairness_obj$get_difference(x_interest, desired_level = "Caucasian", desired_prob = c(0.5,1))
  print(difference)
}


}
\section{Super classes}{
\code{\link[moccf:CounterfactualMethod]{moccf::CounterfactualMethod}} -> \code{\link[moccf:CounterfactualMethodClassif]{moccf::CounterfactualMethodClassif}} -> \code{\link[moccf:MOCClassif]{moccf::MOCClassif}} -> \code{FairnessTest}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FairnessTest$new()}}
\item \href{#method-get_difference}{\code{FairnessTest$get_difference()}}
\item \href{#method-print_prediction}{\code{FairnessTest$print_prediction()}}
\item \href{#method-print_counterfactuals}{\code{FairnessTest$print_counterfactuals()}}
\item \href{#method-prediction_percentages}{\code{FairnessTest$prediction_percentages()}}
\item \href{#method-get_cfactuals_mean}{\code{FairnessTest$get_cfactuals_mean()}}
\item \href{#method-plot_tSNE}{\code{FairnessTest$plot_tSNE()}}
\item \href{#method-clone}{\code{FairnessTest$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="moccf" data-topic="CounterfactualMethod" data-id="print">}\href{../../moccf/html/CounterfactualMethod.html#method-print}{\code{moccf::CounterfactualMethod$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="moccf" data-topic="CounterfactualMethodClassif" data-id="find_counterfactuals">}\href{../../moccf/html/CounterfactualMethodClassif.html#method-find_counterfactuals}{\code{moccf::CounterfactualMethodClassif$find_counterfactuals()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="moccf" data-topic="MOCClassif" data-id="get_dominated_hv">}\href{../../moccf/html/MOCClassif.html#method-get_dominated_hv}{\code{moccf::MOCClassif$get_dominated_hv()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="moccf" data-topic="MOCClassif" data-id="plot_search">}\href{../../moccf/html/MOCClassif.html#method-plot_search}{\code{moccf::MOCClassif$plot_search()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="moccf" data-topic="MOCClassif" data-id="plot_statistics">}\href{../../moccf/html/MOCClassif.html#method-plot_statistics}{\code{moccf::MOCClassif$plot_statistics()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new FairnessTest object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessTest$new(
  predictor = NULL,
  df = NULL,
  sensitive_attribute = NULL,
  epsilon = NULL,
  fixed_features = NULL,
  max_changed = NULL,
  mu = 20L,
  n_generations = 175L,
  p_rec = 0.57,
  p_rec_gen = 0.85,
  p_rec_use_orig = 0.88,
  p_mut = 0.79,
  p_mut_gen = 0.56,
  p_mut_use_orig = 0.32,
  k = 1L,
  weights = NULL,
  lower = NULL,
  upper = NULL,
  init_strategy = "random",
  use_conditional_mutator = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictor}}{(\verb{iml predictor})\cr
The original predictor where as response variable we use \code{Two_yr_Recidivism}.}

\item{\code{df}}{(dataframe)\cr
The dataset or dataframe for the prediction}

\item{\code{sensitive_attribute}}{(string)\cr
The name of sensitive_attribute.}

\item{\code{row_num}}{(numeric())\cr
The row number of instance of interest \code{x_interest}.}

\item{\code{desired_level}}{(string)\cr
The desired class we want to have in our counterfactuals}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_difference"></a>}}
\if{latex}{\out{\hypertarget{method-get_difference}{}}}
\subsection{Method \code{get_difference()}}{
Creates a dataframe with the differences of predictions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessTest$get_difference(x_interest, desired_level, desired_prob)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A dataframe with the differences of predictions of the original instance and the counterfactuals
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_prediction"></a>}}
\if{latex}{\out{\hypertarget{method-print_prediction}{}}}
\subsection{Method \code{print_prediction()}}{
returns a dataframe with the required column for tSNE plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessTest$print_prediction()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A dataframe of counterfactuals with the required column for tSNE plot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_counterfactuals"></a>}}
\if{latex}{\out{\hypertarget{method-print_counterfactuals}{}}}
\subsection{Method \code{print_counterfactuals()}}{
prints the counterfactuals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessTest$print_counterfactuals()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
prints the generated counterfactuals
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prediction_percentages"></a>}}
\if{latex}{\out{\hypertarget{method-prediction_percentages}{}}}
\subsection{Method \code{prediction_percentages()}}{
plots a bar plot for the predictions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessTest$prediction_percentages()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A barplot for the predictions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_cfactuals_mean"></a>}}
\if{latex}{\out{\hypertarget{method-get_cfactuals_mean}{}}}
\subsection{Method \code{get_cfactuals_mean()}}{
returns mean of the prediction of the counterfactuals as the opposite of the x_interest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessTest$get_cfactuals_mean()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
mean of the prediction of the counterfactuals as the opposite of the x_interest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_tSNE"></a>}}
\if{latex}{\out{\hypertarget{method-plot_tSNE}{}}}
\subsection{Method \code{plot_tSNE()}}{
plots tSNE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessTest$plot_tSNE(x_interest)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tSNE plot for the dataset with counterfactuals and x_interest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessTest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
